pxytoggle: switched to using on/off argument
Also the script is now reading the PORT and HOST values for the proxy
from a shared file, which is currently in the same directory in a file
named .proxy_env.  This allows other scripts/utilities to also know what
the state of the proxy's host and port should be, like a script that
sets the http_proxy ENV variable in a Bash session.