vim: setup basic plumbing for LSP config file
The previous lua/lsp-config.lua file was too complicated for me to get
the initial setup working so I decided to simplify things and build up
an LSP config from the ground-up.  I also decided to do it all in a
Vimscript file since I am more familiar with Vimscript and becuase you
can have Lua in a Vimscript file.