git: now allow for global name/email override
Previously the .gitconfig name & email were statically set to my Github
name and email.  This doesn't work for work computers however.

This commit updates things so that the dotfiles and ansible/pc-setup
repos will use my Github name and email but the global (~/.gitconfig)
file can be set to something suitable to the computer the dotfiles repo
is being installed on, like a work email address.