vim: Moved most configs back into .vimrc
Managing my vim config across so many different files was actually more
of a pain then a help.  Now that I better know how to use folds, I
decided to move most of the config back into the .vimrc file so it is
easier to make config changes.  I think I went overly complex before it
was warranted.  If things need to be broken out into separate files in
the future as it makes sense, then I will do so.