alacritty: set TERM=xterm-256color for Nvim
In order to get the nvim-tree colors working properly, you need to set
the vim setting `set termguicolors` on.  However, in Alacritty that setting
didn't seem to work without setting the TERM environment variable to
either 'xterm-256color' (this commit) or 'alacritty'.  I left it as the
former setting since 'xterm-256color' seems more widely applicable.

I don't know why not setting this env var in alacritty.yml causes `set
termguicolors` to change the color scheme, though it seems TERM by
default gets set to `screen-256color`, so it could be that that default
value is the issue.