tmux/alacritty: do real fix for termguicolors
The last commit stated that the TERM variable needed to be set in
alacritty.yml for `termguicolors` to work properly in nvim.  However, I
realized that setting only *seemed* to work b/c I was testing nvim
outside of tmux so it made it seem like it worked.  In fact, it still
didn't fix things inside a tmux session.

This commit adds the proper fix for things by ensuring that truecolor
support is enable from *within* tmux.

I also added a script that will show a smooth color gradient if
truecolors are properly working.  The script, along with the required
.tmux.conf changes, were found here:

  - https://jdhao.github.io/2018/10/19/tmux_nvim_true_color/