nvim: moved all keymaps into single file
It is getting increasingly difficult to manage all the keymaps I have
b/c they are split over so many files based on context.  I decided to
move them all into a single file so I can have better control over them
and help eliminate risk of mapping conflicts.

That said, there are some keymappings setup in Lua scripts in the
after/plugin directory b/c they need to be defined in the context of the
lua (or it would just be really hard to figure out how to have the same
functionality in a .vim file).  Also, the majority of my standard
keymaps are in my vimrc-common.vim file.

At some point, I should probably just decide to move them all to Neovim
or do conditional includes in the keymaps.vim file.