Convert install personal to modular install script
Now that the dotfile install and dependency install tasks have started
to be broken down into remote vs. local tasks, we can develop the
concept of dotfile "installation profiles" where we install/setup the
dotfiles in a different way based on the type of machine we are
configuring.  This commit is converting the install script into one that
can be provided a flag that will tell it which type of install to
complete.