shells: add interactive check in both Bash/ZSH shells
We had been checking for an interactive session in the .config/sh/rc
file, but not explicitly in the Bash/ZSH shells.  This would cause
issues b/c as an example with Bash:

  1. .bashrc would source .config/sh/rc
  2. .config/sh/rc would return immediately b/c session is non-interactive
  3. .bashrc would continue on and then throw errors b/c it needed library
     functions loaded from files sourced in .config/sh/rc after the
     interactive session check occurred

The check was left in .config/sh/rc in case we ever did something with
just a plain /bin/sh shell.